pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kg-drill-extraction"
version = "0.1.0"
description = "地质钻孔实体提取与坐标推理系统"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "KG Team", email = "kg@example.com"},
]
maintainers = [
    {name = "KG Team", email = "kg@example.com"},
]
keywords = [
    "NLP",
    "entity extraction",
    "geological data",
    "LLM",
    "coordinate inference",
    "drilling reports",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]

dependencies = [
    "openai>=1.30.0",
    "python-docx>=0.8.11",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "PyYAML>=6.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "matplotlib>=3.7.0",
    "openpyxl>=3.1.0",
    "aiofiles>=23.0.0",
    "sentencepiece>=0.1.99",
    "tiktoken>=0.5.0",
    "protobuf>=3.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
    "pre-commit>=3.0.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "sphinx-autodoc-typehints>=1.23.0",
]
all = [
    "kg-drill-extraction[dev,docs]",
]

[project.urls]
Homepage = "https://github.com/kg-team/kg-drill-extraction"
Documentation = "https://kg-drill-extraction.readthedocs.io"
Repository = "https://github.com/kg-team/kg-drill-extraction.git"
"Bug Tracker" = "https://github.com/kg-team/kg-drill-extraction/issues"

[project.scripts]
kg-extract = "kg_drill_extraction.cli:main"
kg-experiment = "kg_drill_extraction.experiment.runner:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["kg_drill_extraction*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
kg_drill_extraction = [
    "configs/*.yaml",
    "data/*.csv",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | tests/fixtures
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=kg_drill_extraction",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src/kg_drill_extraction"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if False:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]